Added authorization and roles

+++ src/restaurants/restaurants.controller.ts
@@ -9,14 +9,21 @@ import {
   Query,
   UseInterceptors,
   UploadedFiles,
+  UseGuards,
+  ForbiddenException,
 } from '@nestjs/common';
+import { AuthGuard } from '@nestjs/passport';
 import { FilesInterceptor } from '@nestjs/platform-express';
 
 import { Query as ExpressQuery } from 'express-serve-static-core';
+import { RolesGuard } from '../auth/guards/roles.guard';
+import { CurrentUser } from '../auth/decorators/current-user.decorator';
+import { User } from '../auth/schemas/user.schema';
 import { CreateRestaurantDto } from './dto/create-restaurant.dto';
 import { UpdateRestaurantDto } from './dto/update-restaurant.dto';
 import { RestaurantsService } from './restaurants.service';
 import { Restaurant } from './schemas/restaurant.schema';
+import { Roles } from '../auth/decorators/roles.decorator';
 
 @Controller('restaurants')
 export class RestaurantsController {
@@ -28,11 +35,14 @@ export class RestaurantsController {
   }
 
   @Post()
+  @UseGuards(AuthGuard(), RolesGuard)
+  @Roles('admin', 'user')
   async createRestaurant(
     @Body()
     restaurant: CreateRestaurantDto,
+    @CurrentUser() user: User,
   ): Promise<Restaurant> {
-    return this.restaurantsService.create(restaurant);
+    return this.restaurantsService.create(restaurant, user);
   }
 
   @Get(':id')
@@ -44,23 +54,35 @@ export class RestaurantsController {
   }
 
   @Put(':id')
+  @UseGuards(AuthGuard())
   async updateRestaurant(
     @Param('id')
     id: string,
     @Body()
     restaurant: UpdateRestaurantDto,
+    @CurrentUser() user: User,
   ): Promise<Restaurant> {
-    await this.restaurantsService.findById(id);
+    const res = await this.restaurantsService.findById(id);
+
+    if (res.user.toString() !== user._id.toString()) {
+      throw new ForbiddenException('You can not update this restaurant.');
+    }
 
     return this.restaurantsService.updateById(id, restaurant);
   }
 
   @Delete(':id')
+  @UseGuards(AuthGuard())
   async deleteRestaurant(
     @Param('id')
     id: string,
+    @CurrentUser() user: User,
   ): Promise<{ deleted: Boolean }> {
     const restaurant = await this.restaurantsService.findById(id);
+
+    if (restaurant.user.toString() !== user._id.toString()) {
+      throw new ForbiddenException('You can not delete this restaurant.');
+    }
 
     const isDeleted = await this.restaurantsService.deleteImages(
       restaurant.images,
@@ -80,6 +102,7 @@ export class RestaurantsController {
   }
 
   @Put('upload/:id')
+  @UseGuards(AuthGuard())
   @UseInterceptors(FilesInterceptor('files'))
   async uploadFiles(
     @Param('id') id: string,
